+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix: API methods GetFavoriteAirings() and GetPotentialFavoriteAirings() were returning all airings for keyword favorites.
 * New: Increased possible range for scheduling lookahead to 21 days. The default is still 14 days.
 * Fix: Removed check in Scheduler that was preventing a future airing beyond lookahead from being considered to resolve a conflict.
+* Fix: Fixed Carny not being marked prepped on startup when no agents exist.
 
 ## Version 9.1.1 (2017-05-22)
 * Fix: Fixed a problem with awk parsing in Ubuntu